Refactor code, many parts of the code were in incorrect layers, resolved, had to resolve these issues first, now continuing progress with manual post method and avoiding clashes between manual post and automatic post
