+++ sahha-android/src/main/AndroidManifest.xml
@@ -25,16 +25,16 @@
             android:label="@string/title_activity_sahha_permission"
             android:theme="@style/Theme.AppCompat.Transparent.NoActionBar" />
 
-        <service android:name=".domain.service.DataCollectionService" />
+        <service android:name=".data.service.DataCollectionService" />
 
         <receiver
-            android:name=".domain.receiver.SleepReceiver"
+            android:name=".data.receiver.SleepReceiver"
             android:enabled="true"
             android:exported="true"
             android:permission="android.permission.ACTIVITY_RECOGNITION" />
 
         <receiver
-            android:name=".domain.receiver.AutoStartReceiver"
+            android:name=".data.receiver.AutoStartReceiver"
             android:enabled="true"
             android:exported="true"
             android:permission="android.permission.RECEIVE_BOOT_COMPLETED">
