+++ sahha-android/src/main/java/sdk/sahha/android/common/SahhaReceiversAndListeners.kt
@@ -4,8 +4,8 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import sdk.sahha.android.data.listeners.StepCounterListener
 import sdk.sahha.android.data.listeners.StepDetectorListener
-import sdk.sahha.android.domain.receiver.PhoneScreenStateReceiver
-import sdk.sahha.android.domain.receiver.TimeZoneChangedReceiver
+import sdk.sahha.android.data.receiver.PhoneScreenStateReceiver
+import sdk.sahha.android.data.receiver.TimeZoneChangedReceiver
 
 object SahhaReceiversAndListeners {
     //Receivers
