+++ sahha-android/src/main/java/sdk/sahha/android/data/remote/SahhaApi.kt
@@ -9,10 +9,7 @@ import sdk.sahha.android.data.Constants.APP_SECRET_HEADER
 import sdk.sahha.android.data.Constants.AUTHORIZATION_HEADER
 import sdk.sahha.android.data.remote.dto.DemographicDto
 import sdk.sahha.android.data.remote.dto.StepDto
-import sdk.sahha.android.data.remote.dto.send.DeviceInformationSendDto
-import sdk.sahha.android.data.remote.dto.send.ExternalIdSendDto
-import sdk.sahha.android.data.remote.dto.send.PhoneUsageSendDto
-import sdk.sahha.android.data.remote.dto.send.SleepSendDto
+import sdk.sahha.android.data.remote.dto.send.*
 import sdk.sahha.android.domain.model.analyze.AnalyzeRequest
 import sdk.sahha.android.domain.model.auth.TokenData
 import sdk.sahha.android.source.SahhaDemographic
@@ -38,7 +35,7 @@ interface SahhaApi {
     @POST("oauth/profile/refreshToken")
     suspend fun postRefreshTokenResponse(
         @Header(AUTHORIZATION_HEADER) profileToken: String,
-        @Body refreshToken: String
+        @Body refreshToken: RefreshTokenSendDto
     ): Response<TokenData>
 
     @POST("movement/log")
