+++ sahha-android/src/main/java/sdk/sahha/android/data/remote/dto/send/RefreshTokenSendDto.kt
@@ -1,0 +1,5 @@
+package sdk.sahha.android.data.remote.dto.send
+
+data class RefreshTokenSendDto(
+    val refreshToken: String
+)
