+++ sahha-android/src/main/java/sdk/sahha/android/data/repository/DeviceInfoRepoImpl.kt
@@ -1,9 +1,24 @@
 package sdk.sahha.android.data.repository
 
 import android.os.Build
+import okhttp3.ResponseBody
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+import sdk.sahha.android.common.ResponseCode
+import sdk.sahha.android.common.SahhaErrorLogger
+import sdk.sahha.android.common.SahhaErrors
+import sdk.sahha.android.data.remote.SahhaApi
+import sdk.sahha.android.domain.model.device_info.DeviceInformation
+import sdk.sahha.android.domain.repository.AuthRepo
 import sdk.sahha.android.domain.repository.DeviceInfoRepo
+import sdk.sahha.android.source.SahhaConverterUtility
 
-class DeviceInfoRepoImpl : DeviceInfoRepo {
+class DeviceInfoRepoImpl(
+    private val authRepo: AuthRepo,
+    private val api: SahhaApi,
+    private val sahhaErrorLogger: SahhaErrorLogger
+) : DeviceInfoRepo {
     override fun getPlatform(): String {
         return "Android"
     }
@@ -18,5 +33,35 @@ class DeviceInfoRepoImpl : DeviceInfoRepo {
 
     override fun getSdkVersion(): String {
         TODO("Not yet implemented")
+    }
+
+    override suspend fun putDeviceInformation(deviceInformation: DeviceInformation) {
+        val call = putDeviceInformationResponse(deviceInformation)
+        call.enqueue(object : Callback<ResponseBody> {
+            override fun onResponse(call: Call<ResponseBody>, response: Response<ResponseBody>) {
+                if (ResponseCode.isSuccessful(response.code())) {
+                    return
+                }
+
+                sahhaErrorLogger.api(call, response)
+            }
+
+            override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
+                sahhaErrorLogger.api(
+                    call,
+                    SahhaErrors.typeResponse,
+                    null,
+                    t.message ?: SahhaErrors.responseFailure
+                )
+            }
+        })
+    }
+
+    private suspend fun putDeviceInformationResponse(deviceInformation: DeviceInformation): Call<ResponseBody> {
+        val token = authRepo.getToken()!!
+        return api.putDeviceInformation(
+            token,
+            SahhaConverterUtility.deviceInfoToDeviceInfoSendDto(deviceInformation)
+        )
     }
 }
