+++ sahha-android/src/main/java/sdk/sahha/android/domain/manager/ReceiverManager.kt
@@ -1,0 +1,11 @@
+package sdk.sahha.android.domain.manager
+
+import android.content.Context
+
+interface ReceiverManager {
+    fun startActivityRecognitionReceiver(callback: ((error: String?, success: Boolean) -> Unit)? = null)
+    fun startPhoneScreenReceivers(
+        serviceContext: Context,
+    )
+    fun startTimeZoneChangedReceiver(context: Context)
+}
