+++ sahha-android/src/main/java/sdk/sahha/android/domain/repository/DeviceInfoRepo.kt
@@ -1,8 +1,11 @@
 package sdk.sahha.android.domain.repository
 
+import sdk.sahha.android.domain.model.device_info.DeviceInformation
+
 interface DeviceInfoRepo {
     fun getPlatform(): String
     fun getPlatformVer(): String
     fun getDeviceModel(): String
     fun getSdkVersion(): String
+    suspend fun putDeviceInformation(deviceInformation: DeviceInformation)
 }
