+++ sahha-android/src/main/java/sdk/sahha/android/domain/repository/UserDataRepo.kt
@@ -1,0 +1,17 @@
+package sdk.sahha.android.domain.repository
+
+import sdk.sahha.android.source.SahhaDemographic
+
+interface UserDataRepo {
+    suspend fun getAnalysis(
+        dates: Pair<String, String>? = null,
+        includeSourceData: Boolean,
+        callback: ((error: String?, successful: String?) -> Unit)?,
+    )
+
+    suspend fun getDemographic(callback: ((error: String?, demographic: SahhaDemographic?) -> Unit)?)
+    suspend fun postDemographic(
+        sahhaDemographic: SahhaDemographic,
+        callback: ((error: String?, successful: Boolean) -> Unit)?
+    )
+}
