+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/GetDemographicUseCase.kt
@@ -1,10 +1,11 @@
 package sdk.sahha.android.domain.use_case
 
-import sdk.sahha.android.domain.repository.RemoteRepo
+import sdk.sahha.android.domain.repository.SensorRepo
+import sdk.sahha.android.domain.repository.UserDataRepo
 import sdk.sahha.android.source.SahhaDemographic
 
 class GetDemographicUseCase (
-    private val repository: RemoteRepo
+    private val repository: UserDataRepo
 ) {
     suspend operator fun invoke(
         callback: ((error: String?, demographic: SahhaDemographic?) -> Unit)?
