+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/GetSensorDataUseCase.kt
@@ -1,9 +1,9 @@
 package sdk.sahha.android.domain.use_case
 
-import sdk.sahha.android.domain.repository.BackgroundRepo
+import sdk.sahha.android.domain.repository.SensorRepo
 import sdk.sahha.android.source.SahhaSensor
 
-class GetSensorDataUseCase(private val repository: BackgroundRepo) {
+class GetSensorDataUseCase(private val repository: SensorRepo) {
     suspend operator fun invoke(
         sensor: SahhaSensor,
         callback: ((error: String?, success: String?) -> Unit)
