+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/background/StartCollectingSleepDataUseCase.kt
@@ -1,10 +1,10 @@
 package sdk.sahha.android.domain.use_case.background
 
 import sdk.sahha.android.data.Constants
-import sdk.sahha.android.domain.repository.BackgroundRepo
+import sdk.sahha.android.domain.repository.SensorRepo
 
 class StartCollectingSleepDataUseCase (
-    private val repository: BackgroundRepo
+    private val repository: SensorRepo
 ) {
     operator fun invoke() {
         repository.startSleepWorker(
