+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/background/StartCollectingStepCounterData.kt
@@ -2,10 +2,10 @@ package sdk.sahha.android.domain.use_case.background
 
 import android.content.Context
 import sdk.sahha.android.data.local.dao.MovementDao
-import sdk.sahha.android.domain.repository.BackgroundRepo
+import sdk.sahha.android.domain.repository.SensorRepo
 
 class StartCollectingStepCounterData (
-    private val repository: BackgroundRepo
+    private val repository: SensorRepo
 ) {
     suspend operator fun invoke(
         context: Context,
