+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/background/StartDataCollectionServiceUseCase.kt
@@ -1,11 +1,16 @@
 package sdk.sahha.android.domain.use_case.background
 
-import sdk.sahha.android.domain.repository.BackgroundRepo
+import sdk.sahha.android.domain.manager.SahhaNotificationManager
 
-class StartDataCollectionServiceUseCase (
-    private val repository: BackgroundRepo
+class StartDataCollectionServiceUseCase(
+    private val manager: SahhaNotificationManager
 ) {
-    operator fun invoke(icon: Int?, title: String?, shortDescription: String?, callback: ((error: String?, success: Boolean) -> Unit)?) {
-        repository.startDataCollectionService(icon, title, shortDescription, callback)
+    operator fun invoke(
+        icon: Int?,
+        title: String?,
+        shortDescription: String?,
+        callback: ((error: String?, success: Boolean) -> Unit)?
+    ) {
+        manager.startDataCollectionService(icon, title, shortDescription, callback)
     }
 }
