+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/permissions/OpenAppSettingsUseCase.kt
@@ -1,10 +1,10 @@
 package sdk.sahha.android.domain.use_case.permissions
 
 import android.content.Context
-import sdk.sahha.android.domain.repository.PermissionsRepo
+import sdk.sahha.android.domain.manager.PermissionManager
 
 class OpenAppSettingsUseCase (
-    private val repository: PermissionsRepo
+    private val repository: PermissionManager
 ) {
     operator fun invoke(context: Context) {
         repository.openAppSettings(context)
