+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/permissions/SetPermissionLogicUseCase.kt
@@ -1,10 +1,10 @@
 package sdk.sahha.android.domain.use_case.permissions
 
 import androidx.activity.ComponentActivity
-import sdk.sahha.android.domain.repository.PermissionsRepo
+import sdk.sahha.android.domain.manager.PermissionManager
 
 class SetPermissionLogicUseCase (
-    private val repository: PermissionsRepo
+    private val repository: PermissionManager
 ) {
     operator fun invoke(activity: ComponentActivity) {
         repository.setPermissionLogic(activity)
