+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/post/PostAllSensorDataUseCase.kt
@@ -1,10 +1,10 @@
 package sdk.sahha.android.domain.use_case.post
 
-import sdk.sahha.android.domain.repository.RemoteRepo
+import sdk.sahha.android.domain.repository.SensorRepo
 import sdk.sahha.android.source.SahhaSensor
 
 class PostAllSensorDataUseCase (
-    val repository: RemoteRepo
+    val repository: SensorRepo
 ) {
     suspend operator fun invoke(
         callback: ((error: String?, success: Boolean) -> Unit)
