+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/post/PostDemographicUseCase.kt
@@ -1,10 +1,10 @@
 package sdk.sahha.android.domain.use_case.post
 
-import sdk.sahha.android.domain.repository.RemoteRepo
+import sdk.sahha.android.domain.repository.UserDataRepo
 import sdk.sahha.android.source.SahhaDemographic
 
-class PostDemographicUseCase (
-    private val repository: RemoteRepo
+class PostDemographicUseCase(
+    private val repository: UserDataRepo
 ) {
     suspend operator fun invoke(
         sahhaDemographic: SahhaDemographic,
