+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/post/PostDeviceDataUseCase.kt
@@ -1,9 +1,9 @@
 package sdk.sahha.android.domain.use_case.post
 
-import sdk.sahha.android.domain.repository.RemoteRepo
+import sdk.sahha.android.domain.repository.SensorRepo
 
 class PostDeviceDataUseCase (
-    private val repository: RemoteRepo
+    private val repository: SensorRepo
 ) {
     suspend operator fun invoke(callback: ((error: String?, success: Boolean) -> Unit)?) {
         repository.postPhoneScreenLockData(callback)
