+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/post/PostStepDataUseCase.kt
@@ -1,10 +1,10 @@
 package sdk.sahha.android.domain.use_case.post
 
 import sdk.sahha.android.domain.model.steps.StepData
-import sdk.sahha.android.domain.repository.RemoteRepo
+import sdk.sahha.android.domain.repository.SensorRepo
 
 class PostStepDataUseCase (
-    val repository: RemoteRepo
+    val repository: SensorRepo
 ) {
     suspend operator fun invoke(
         stepData: List<StepData>,
