+++ sahha-android/src/main/java/sdk/sahha/android/domain/use_case/post/StartPostWorkersUseCase.kt
@@ -1,9 +1,9 @@
 package sdk.sahha.android.domain.use_case.post
 
-import sdk.sahha.android.domain.repository.BackgroundRepo
+import sdk.sahha.android.domain.repository.SensorRepo
 
 class StartPostWorkersUseCase (
-    private val repository: BackgroundRepo
+    private val repository: SensorRepo
 ) {
     operator fun invoke() {
         repository.startPostWorkersAsync()
