+++ sahha-android/src/main/java/sdk/sahha/android/source/SahhaSettings.kt
@@ -8,5 +8,4 @@ class SahhaSettings(
     val notificationSettings: SahhaNotificationConfiguration? = null,
     val framework: SahhaFramework = SahhaFramework.android_kotlin,
     val sensors: Set<Enum<SahhaSensor>>? = null,
-    val postSensorDataManually: Boolean = false
 )
